+++ ChangeLog.md
@@ -7,3 +7,6 @@
 - Added docs to `monitor.Timer`
 - Fixed `get_file_and_stream_logger` and `get_stream_logger` functions in `log.py`.
 - Bump version to `0.2.0`.
+
+## [2021-05-18 Tue 05:54]
+- Added `apply_list` and `set_lens`.